Add details to Logic section of DG
The Logic section lacked clarity in explaining the way in which command
execution occurred in the Logic component. Thus, I have added several
more sequence diagrams and ref frames to explain it in greater detail,
that is also maintainable as I have generalised and made it extensible
when more types of commands are added.