Enhance JSON File Reading Verification
Users are allow to directly make changes
to the JSON file that stores the data of
the contact list.

Changes made that violates the affiliation
rules are not checked, and is passed into
the contact list haphazardly.

Let's enhance the way the JSON file is read,
so that illegal changes are flagged.