Add array option for validity checks, tests
Previously, the validity checks for each attribute (e.g. isValidName,
isValidPhone etc) could only be applied to single strings.

Let's use method overloading to allow the method to take in an array as
well, and check for validity of every element in the array of strings.
In addition, tests were added for FindCommandParser.